+++ StereoPlus2/Settings.hpp
@@ -63,15 +63,6 @@ public:
 	StaticProperty(float, CustomRenderWindowAlpha)
 
 
-	// Shortcuts
-	StaticProperty(Key::Combination, TransformToolShortcut)
-	StaticProperty(Key::Combination, PenToolShortcut)
-	StaticProperty(Key::Combination, ExtrusionToolShortcut)
-	StaticProperty(Key::Combination, RenderViewportToFile)
-	StaticProperty(Key::Combination, RenderAdvancedToFile)
-
-	StaticProperty(Key::Combination, SwitchUseDiscreteMovement)
-
 	static const std::string& Name(void* reference) {
 		static std::map<void*, const std::string> v = {
 			{&Language,"language"},