+++ StereoPlus2/Windows.hpp
@@ -601,10 +601,6 @@ public:
 	}
 };
 
-
-
-
-template<ObjectType type>
 class PointPenToolWindow : Window, Attributes {
 	const Log log = Log::For<SceneObjectInspectorWindow>();
 
@@ -651,33 +647,32 @@ class PointPenToolWindow : Window, Attributes {
 	}
 
 	bool DesignInternal() {
-		ImGui::Text(GetName(type, GetTarget()).c_str());
+		//ImGui::Text(GetName(type, GetTarget()).c_str());
+		//if (ImGui::BeginDragDropTarget())
+		//{
+		//	ImGuiDragDropFlags target_flags = 0;
+		//	//target_flags |= ImGuiDragDropFlags_AcceptBeforeDelivery;    // Don't wait until the delivery (release mouse button on a target) to do something
+		//	//target_flags |= ImGuiDragDropFlags_AcceptNoDrawDefaultRect; // Don't display the yellow rectangle
+		//	std::vector<PON> objects;
+		//	if (DragDropBuffer::PopDragDropPayload("SceneObjects", target_flags, &objects))
+		//	{
+		//		if (objects.size() > 1) {
+		//			log.Warning("Drawing instrument can't accept multiple scene objects");
+		//		}
+		//		else {
+		//			if (!tool->BindSceneObjects(objects))
+		//				return false;
+		//		}
+		//	}
 
-		if (ImGui::BeginDragDropTarget())
-		{
-			ImGuiDragDropFlags target_flags = 0;
-			//target_flags |= ImGuiDragDropFlags_AcceptBeforeDelivery;    // Don't wait until the delivery (release mouse button on a target) to do something
-			//target_flags |= ImGuiDragDropFlags_AcceptNoDrawDefaultRect; // Don't display the yellow rectangle
-			std::vector<PON> objects;
-			if (DragDropBuffer::PopDragDropPayload("SceneObjects", target_flags, &objects))
-			{
-				if (objects.size() > 1) {
-					log.Warning("Drawing instrument can't accept multiple scene objects");
-				}
-				else {
-					if (!tool->BindSceneObjects(objects))
-						return false;
-				}
-			}
-
-			ImGui::EndDragDropTarget();
-		}
+		//	ImGui::EndDragDropTarget();
+		//}
 
 		{
 			bool isActive = (GetTarget()) != nullptr;
 			if (IsActive(isActive))
 			{
-				if (ImGui::Button("Release"))
+				if (ImGui::Button(isActive ? "Release" : "No objects bind"))
 				{
 					tool->UnbindSceneObjects();
 				}
@@ -701,7 +696,7 @@ public:
 	// If this is null then the window probably wasn't initialized.
 	//SceneObject* target = nullptr;
 
-	PointPenEditingTool<type>* tool = nullptr;
+	PointPenEditingTool* tool = nullptr;
 
 	virtual SceneObject* GetTarget() {
 		if (tool == nullptr)
@@ -1111,8 +1106,7 @@ class ToolWindow : Window {
 
 	template<typename T>
 	void ConfigureCreationTool(CreatingTool<T>& creatingTool, std::function<void(SceneObject*)> initFunc) {
-		creatingTool.scene.BindAndApply(scene);
-		creatingTool.destination.BindAndApply(scene.Get()->root);
+		creatingTool.destination.BindAndApply(Scene::root());
 		creatingTool.init = initFunc;
 	}
 
@@ -1200,13 +1194,13 @@ public:
 		if (ImGui::Button(LocaleProvider::GetC("tool:extrusion")))
 			ApplyTool<ExtrusionToolWindow<StereoPolyLineT>, ExtrusionEditingTool<StereoPolyLineT>>();
 		if (ImGui::Button(LocaleProvider::GetC("tool:pen")))
-			ApplyTool<PointPenToolWindow<StereoPolyLineT>, PointPenEditingTool<StereoPolyLineT>>();
+			ApplyTool<PointPenToolWindow, PointPenEditingTool>();
 		if (ImGui::Button(LocaleProvider::GetC("tool:transformation")))
 			ApplyTool<TransformToolWindow, TransformTool>();
 
 		{
 			ImGui::Separator();
-			static int v = (int)Settings::SpaceMode().Get();
+			auto v = (int)Settings::SpaceMode().Get();
 			if (ImGui::RadioButton(LocaleProvider::GetC("world"), &v, (int)SpaceMode::World))
 				Settings::SpaceMode() = SpaceMode::World;
 			if (ImGui::RadioButton(LocaleProvider::GetC("local"), &v, (int)SpaceMode::Local))