+++ CHANGE_LOG.md
@@ -22,5 +22,7 @@
 - reworked control. Now to use mouse control - drag scene image. Also, depending on the desired action (translate, rotate, scale) - use (Alt, Ctrl, Shift) respectively;
 - objects can be rotated around 2 axes simultaneously;
 - enabled translation in local space for multiple objects. They are being moved in respect to cross rotation;
+- fixed ObjectSelection sending 2 events on Set();
+- fixed IsDown/IsPressed(<key>, false) methods working the other way around;
+- added shortcut Ctrl+D for deselect all;
 
-! in windows, Shift is overlapping with numpad keys so combinations Shift+Numpad aren't possible https://github.com/glfw/glfw/issues/670